extend-stages.md - Rename "Maintained/Unmaintained" to "Stable/Deprecated"
During the meeting of the Extension Working group on 27-Oct-16, we discussed
trade-offs in the names.  (Ex: Now that "Stewardship" has an option for
"Seeking Maintainer", it seems a bit confusing to have "Maturity" as
"Maintained/Unmaintained".) There was support changing these two names and
no opposition.

The goal is to publicize this soon in a public blog.  Hopefully, this is the
last round, but if there's strong feedback we could revise again.