hook_civicrm_* - Fix various parameter names
This updates the docs to ensure that hook parameters exactly match the names
in the code.

Prior to 4.7.19, the names of hook-parameters didn't really matter (as long
as they were in the right order).  In 4.7.19, the names took on extra
significance because Symfony-based listeners use objects and fields (e.g.
`$event->firstParam`) instead of ordered params (`function($firstParam)`).