debugging.md - Remove misleading link
The doc mentions the page "Administer => System Settings => Cleanup Caches and Update Paths" as a way to clear caches. This is half-true: it does clear `templates_c` and the `civicrm_cache` table, but there are other things that it doesn't touch. (Compare `CRM_Admin_form_Setting_UpdateConfigBackend::postProcess` with the `CRM_Core_Invoke::rebuildMenuAndCaches`.)

IMHO, `UpdateConfigBackend::postProcess` should be patched to just do a general system flush (`civicrm_api3('System', 'flush', array())`) -- but (for now) the page is misleading and we shouldn't direct people there.