CRM-20926 - hook_civicrm_xmlMenu - Fill in some missing docs on XML schema
To document the changes in CRM-20926, we need some kind of doc about the XML
menu structure...  which doesn't seem to exist.

This change fixes some quirks in the `h_c_xmlMenu` doc and also adds a
section for the XML.  This revision is not exhausive, but it covers some
common ones, and it's enough that we can add-in more.