Move CiviReport page into "framework" chapter
Why? Because right now there's where lots of other content exists which
is pretty similar to this page. And there's no "reference" chapter. I
think that once we're finished with the wiki migration we might like to
clean up that "framework" chapter a bit by moving some pages around. But
until then, I think it's easiest if we keep content together which is
similar.