Improve docs for publishing an extension.
Previously it read that you only needed to create a release on *subsequent* releases, but my experience was that the extension was not available until the initial release was made also. Rewrote to clarify this.

Also made a few additions to inline docs on the create a release node page. These can be recorded over in CiviCRM.org website repo via Features.

Refs #323.