change path in Entity creation w/flow
I'm not sure if this has only come about because of my ineptitude (i.e. whether an Entity would always have the same name as the extension), but I found that I couldn't find the files referred to because my Entity and my Extension had different names.

Would that be bad practice on my part? (it would be good to know as I'm very new to coding)

Either way, I think this may be more accurate... rest is up to you.

Cheers,
M.