Clarify requirements for publication in ext dir
I removed this sentence because I don't think it's accurate.

We have these review criteria:
https://docs.civicrm.org/extensions/lifecycle/#review-criteria

The criteria do say that you can't use PHP overrides. But there are
other interfaces (like BAO functions for example) that are not in the
"supported methods" list but also not prohibited based on the extension
review criteria.