buildkit.md - Move full OS table to appendix
This table is not relevant for several install methods (Vagrant/Docker/Generic).

For Ubuntu/Debian, I think it's useful information, but it's a bit
overwhelming, and what we're basically doing is elaborating on a couple
words (*it needs to be a __recent__, supported version of Ubuntu/Debian -
which is specifically yadda yadda...*).

If one is oblivious to this content and runs the install script on an
unsupported version of Ubuntu/Debian, that should be safe -- since the first
thing it does is to verify compatibility.

Note: not trying to bury - the next commit will call this out a bit more.