tools/issue-tracking.md - Update guidance on Gitlab labels
__Before__: The documentation explains ~5 common labels.  Gitlab actually has ~50
labels.

__After__: The documentation explains ~3 important labels and ~4 *prefixes*.
Each prefix is given with a definition and example, and there are links
to the full lists.

__Comment__: The link goes to https://lab.civicrm.org/dev/core/-/labels
though arguably https://lab.civicrm.org/groups/dev/-/labels might be better.
However, there a dozen or so labels which (rightly or wrongly) only show up
in `dev/core`, so `dev/core` gives the more comprehensive listing.